+++ module-dvbapi.h
@@ -49,55 +49,55 @@ struct box_devices
 
 struct s_ecmpids
 {
-	unsigned short CAID;
-	unsigned long PROVID;
-	unsigned short ECM_PID;
-	unsigned short EMM_PID;
-	int irdeto_numchids;
-	int irdeto_curchid;
-	int irdeto_chid;
-	int checked;
-	int status;
+	uint16_t CAID;
+	uint32_t PROVID;
+	uint16_t ECM_PID;
+	uint16_t EMM_PID;
+	int32_t irdeto_numchids;
+	int32_t irdeto_curchid;
+	int32_t irdeto_chid;
+	int32_t checked;
+	int32_t status;
 	unsigned char table;
-	int index;
+	int32_t index;
 	uint32_t streams;
 };
 
 typedef struct filter_s
 {
-	uint fd; //FilterHandle
-	int pidindex;
-	int pid;
-	ushort type;
-	int count;
+	uint32_t fd; //FilterHandle
+	int32_t pidindex;
+	int32_t pid;
+	uint16_t type;
+	int32_t count;
 #ifdef WITH_STAPI
-	int NumSlots;
-	uint	SlotHandle[10];
-	uint  	BufferHandle[10];
+	int32_t NumSlots;
+	uint32_t	SlotHandle[10];
+	uint32_t  	BufferHandle[10];
 #endif
 } FILTERTYPE;
 
 struct s_emmpids
 {
-	ushort CAID;
-	ulong PROVID;
-	ushort PID;
-	uint8 type;
+	uint16_t CAID;
+	uint32_t PROVID;
+	uint16_t PID;
+	uint8_t type;
 };
 
 #ifdef WITH_STAPI
 struct STDEVICE
 {
 	char name[20];
-	uint 	SessionHandle;
-	uint	SignalHandle;
+	uint32_t 	SessionHandle;
+	uint32_t	SignalHandle;
 	pthread_t thread;
 	struct filter_s demux_fd[MAX_DEMUX][MAX_FILTER];
 };
 
 struct read_thread_param
 {
-	int id;
+	int32_t id;
 	struct s_client *cli;
 };
 
@@ -113,51 +113,51 @@ struct STDEVICE dev_list[PTINUM];
 
 typedef struct demux_s
 {
-	int demux_index;
+	int32_t demux_index;
 	FILTERTYPE demux_fd[MAX_FILTER];
-	int ca_mask;
-	int adapter_index;
-	int socket_fd;
-	int ECMpidcount;
+	int32_t ca_mask;
+	int32_t adapter_index;
+	int32_t socket_fd;
+	int32_t ECMpidcount;
 	struct s_ecmpids ECMpids[ECM_PIDS];
-	int EMMpidcount;
+	int32_t EMMpidcount;
 	struct s_emmpids EMMpids[ECM_PIDS];
-	int STREAMpidcount;
-	unsigned short STREAMpids[ECM_PIDS];
-	int pidindex;
-	int curindex;
-	int tries;
-	int max_status;
-	unsigned short program_number;
+	int32_t STREAMpidcount;
+	uint16_t STREAMpids[ECM_PIDS];
+	int32_t pidindex;
+	int32_t curindex;
+	int32_t tries;
+	int32_t max_status;
+	uint16_t program_number;
 	unsigned char lastcw[2][8];
-	int emm_filter;
+	int32_t emm_filter;
 	uchar hexserial[8];
 	struct s_reader *rdr;
 	char pmt_file[30];
-	int pmt_time;
+	int32_t pmt_time;
 #ifdef WITH_STAPI
-	uint DescramblerHandle[PTINUM];
-	int desc_pidcount;
-	uint slot_assc[PTINUM][SLOTNUM];
+	uint32_t DescramblerHandle[PTINUM];
+	int32_t desc_pidcount;
+	uint32_t slot_assc[PTINUM][SLOTNUM];
 #endif
 } DEMUXTYPE;
 
 struct s_dvbapi_priority
 {
 	char type; // p or i
-	ushort caid;
-	ulong provid;
-	ushort srvid;
-	ushort chid;
-	ushort ecmpid;
-	ushort mapcaid;
-	ulong mapprovid;
-	int delay;
-	int force;
+	uint16_t caid;
+	uint32_t provid;
+	uint16_t srvid;
+	uint16_t chid;
+	uint16_t ecmpid;
+	uint16_t mapcaid;
+	uint32_t mapprovid;
+	int32_t delay;
+	int32_t force;
 #ifdef WITH_STAPI
 	char devname[30];
 	char pmtfile[30];
-	int disablefilter;
+	int32_t disablefilter;
 #endif
 	struct s_dvbapi_priority *next;
 };
@@ -214,14 +214,14 @@ struct dmx_sct_filter_params
 };
 
 typedef struct ca_descr {
-	unsigned int index;
-	unsigned int parity;	/* 0 == even, 1 == odd */
+	uint32_t index;
+	uint32_t parity;	/* 0 == even, 1 == odd */
 	unsigned char cw[8];
 } ca_descr_t;
 
 typedef struct ca_pid {
-	unsigned int pid;
-	int index;		/* -1 == disable*/
+	uint32_t pid;
+	int32_t index;		/* -1 == disable*/
 } ca_pid_t;
 
 #define DMX_START		_IO('o', 41)
@@ -236,69 +236,69 @@ typedef struct ca_pid {
 #include "openxcas/openxcas_api.h"
 #include "openxcas/openxcas_message.h"
 
-int openxcas_provid, openxcas_seq, openxcas_filter_idx, openxcas_stream_id, openxcas_cipher_idx, openxcas_busy;
+int32_t openxcas_provid, openxcas_seq, openxcas_filter_idx, openxcas_stream_id, openxcas_cipher_idx, openxcas_busy;
 unsigned char openxcas_cw[16];
-unsigned short openxcas_sid, openxcas_caid, openxcas_ecm_pid, openxcas_video_pid, openxcas_audio_pid, openxcas_data_pid;
+uint16_t openxcas_sid, openxcas_caid, openxcas_ecm_pid, openxcas_video_pid, openxcas_audio_pid, openxcas_data_pid;
 
-void azbox_openxcas_ecm_callback(int stream_id, unsigned int sequence, int cipher_index, unsigned int caid, unsigned char *ecm_data, int l, unsigned short pid);
-void azbox_openxcas_ex_callback(int stream_id, unsigned int seq, int idx, unsigned int pid, unsigned char *ecm_data, int l);
+void azbox_openxcas_ecm_callback(int32_t stream_id, uint32_t sequence, int32_t cipher_index, uint32_t caid, unsigned char *ecm_data, int32_t l, uint16_t pid);
+void azbox_openxcas_ex_callback(int32_t stream_id, uint32_t seq, int32_t idx, uint32_t pid, unsigned char *ecm_data, int32_t l);
 void azbox_send_dcw(struct s_client *client, ECM_REQUEST *er);
 void * azbox_main(void * cli);
 #endif
 
 #ifdef COOL
-int coolapi_set_filter (int fd, int num, int pid, byte * flt, byte * mask);
-int coolapi_remove_filter (int fd, int num);
-int coolapi_open_device (int demux_index, int demux_id);
-int coolapi_close_device(int fd);
-int coolapi_write_cw(int mask, unsigned short *STREAMpids, int count, ca_descr_t * ca_descr);
-int coolapi_set_pid (int demux_id, int num, int index, int pid);
+int32_t coolapi_set_filter (int32_t fd, int32_t num, int32_t pid, byte * flt, byte * mask);
+int32_t coolapi_remove_filter (int32_t fd, int32_t num);
+int32_t coolapi_open_device (int32_t demux_index, int32_t demux_id);
+int32_t coolapi_close_device(int32_t fd);
+int32_t coolapi_write_cw(int32_t mask, uint16_t *STREAMpids, int32_t count, ca_descr_t * ca_descr);
+int32_t coolapi_set_pid (int32_t demux_id, int32_t num, int32_t index, int32_t pid);
 void coolapi_close_all();
-void dvbapi_write_cw(int demux_id, uchar *cw, int index);
+void dvbapi_write_cw(int32_t demux_id, uchar *cw, int32_t index);
 #endif
 
 #ifdef WITH_STAPI
-static int stapi_open();
-static int stapi_set_filter(int demux_id, ushort pid, uchar *filter, uchar *mask, int num, char *pmtfile);
-static int stapi_remove_filter(int demux_id, int num, char *pmtfile);
-static int stapi_set_pid(int demux_id, int num, int index, ushort pid, char *pmtfile);
-static int stapi_write_cw(int demux_id, uchar *cw, ushort *, int, char *pmtfile);
-static int stapi_do_set_filter(int demux_id, FILTERTYPE *filter, ushort *pids, int pidcount, uchar *filt, uchar *mask, int dev_id);
-static int stapi_do_remove_filter(int demux_id, FILTERTYPE *filter, int dev_id);
+static int32_t stapi_open();
+static int32_t stapi_set_filter(int32_t demux_id, uint16_t pid, uchar *filter, uchar *mask, int32_t num, char *pmtfile);
+static int32_t stapi_remove_filter(int32_t demux_id, int32_t num, char *pmtfile);
+static int32_t stapi_set_pid(int32_t demux_id, int32_t num, int32_t index, uint16_t pid, char *pmtfile);
+static int32_t stapi_write_cw(int32_t demux_id, uchar *cw, uint16_t *, int, char *pmtfile);
+static int32_t stapi_do_set_filter(int32_t demux_id, FILTERTYPE *filter, uint16_t *pids, int32_t pidcount, uchar *filt, uchar *mask, int32_t dev_id);
+static int32_t stapi_do_remove_filter(int32_t demux_id, FILTERTYPE *filter, int32_t dev_id);
 static void *stapi_read_thread(void *);
 
-uint oscam_stapi_Capability(char *name);
+uint32_t oscam_stapi_Capability(char *name);
 char *oscam_stapi_LibVersion(void);
-uint oscam_stapi_Open(char *name, uint *sessionhandle);
-uint oscam_stapi_SignalAllocate(uint sessionhandle, uint *signalhandle);
-uint oscam_stapi_FilterAllocate(uint sessionhandle, uint *filterhandle);
-uint oscam_stapi_SlotInit(uint sessionhandle, uint signalhandle, uint *bufferhandle, uint *slothandle, ushort pid);
-uint oscam_stapi_FilterSet(uint filterhandle, uchar *filt, uchar *mask);
-uint oscam_stapi_FilterAssociate(uint filterhandle, uint slothandle);
-uint oscam_stapi_SlotDeallocate(uint slothandle);
-uint oscam_stapi_BufferDeallocate(uint bufferhandle);
-uint oscam_stapi_FilterDeallocate(uint filterhandle);
-uint oscam_stapi_Close(uint sessionhandle);
-uint oscam_stapi_CheckVersion();
-uint oscam_stapi_DescramblerAssociate(uint deschandle, uint slot);
-uint oscam_stapi_DescramblerDisassociate(uint deschandle, uint slot);
-uint oscam_stapi_DescramblerAllocate(uint sessionhandle, uint *deschandle);
-uint oscam_stapi_DescramblerDeallocate(uint deschandle);
-uint oscam_stapi_DescramblerSet(uint deschandle, int parity, uchar *cw);
-uint oscam_stapi_SignalWaitBuffer(uint signalhandle, uint *qbuffer, int timeout);
-uint oscam_stapi_BufferReadSection(uint bufferhandle, uint *filterlist, int maxfilter, uint *filtercount, int *crc, uchar *buf, int bufsize, uint *size);
-uint oscam_stapi_SignalAbort(uint signalhandle);
-uint oscam_stapi_PidQuery(char *name, ushort pid);
-uint oscam_stapi_BufferFlush(uint bufferhandle);
-uint oscam_stapi_SlotClearPid(uint slot);
+uint32_t oscam_stapi_Open(char *name, uint32_t *sessionhandle);
+uint32_t oscam_stapi_SignalAllocate(uint32_t sessionhandle, uint32_t *signalhandle);
+uint32_t oscam_stapi_FilterAllocate(uint32_t sessionhandle, uint32_t *filterhandle);
+uint32_t oscam_stapi_SlotInit(uint32_t sessionhandle, uint32_t signalhandle, uint32_t *bufferhandle, uint32_t *slothandle, uint16_t pid);
+uint32_t oscam_stapi_FilterSet(uint32_t filterhandle, uchar *filt, uchar *mask);
+uint32_t oscam_stapi_FilterAssociate(uint32_t filterhandle, uint32_t slothandle);
+uint32_t oscam_stapi_SlotDeallocate(uint32_t slothandle);
+uint32_t oscam_stapi_BufferDeallocate(uint32_t bufferhandle);
+uint32_t oscam_stapi_FilterDeallocate(uint32_t filterhandle);
+uint32_t oscam_stapi_Close(uint32_t sessionhandle);
+uint32_t oscam_stapi_CheckVersion();
+uint32_t oscam_stapi_DescramblerAssociate(uint32_t deschandle, uint32_t slot);
+uint32_t oscam_stapi_DescramblerDisassociate(uint32_t deschandle, uint32_t slot);
+uint32_t oscam_stapi_DescramblerAllocate(uint32_t sessionhandle, uint32_t *deschandle);
+uint32_t oscam_stapi_DescramblerDeallocate(uint32_t deschandle);
+uint32_t oscam_stapi_DescramblerSet(uint32_t deschandle, int32_t parity, uchar *cw);
+uint32_t oscam_stapi_SignalWaitBuffer(uint32_t signalhandle, uint32_t *qbuffer, int32_t timeout);
+uint32_t oscam_stapi_BufferReadSection(uint32_t bufferhandle, uint32_t *filterlist, int32_t maxfilter, uint32_t *filtercount, int32_t *crc, uchar *buf, int32_t bufsize, uint32_t *size);
+uint32_t oscam_stapi_SignalAbort(uint32_t signalhandle);
+uint32_t oscam_stapi_PidQuery(char *name, uint16_t pid);
+uint32_t oscam_stapi_BufferFlush(uint32_t bufferhandle);
+uint32_t oscam_stapi_SlotClearPid(uint32_t slot);
 #endif
 
 void dvbapi_stop_descrambling(int);
-void dvbapi_process_input(int demux_id, int filter_num, uchar *buffer, int len);
-int dvbapi_open_device(int, int, int);
-int dvbapi_stop_filternum(int demux_index, int num);
-int dvbapi_stop_filter(int demux_index, int type);
-struct s_dvbapi_priority *dvbapi_check_prio_match(int demux_id, int pidindex, char type);
+void dvbapi_process_input(int32_t demux_id, int32_t filter_num, uchar *buffer, int32_t len);
+int32_t dvbapi_open_device(int, int, int);
+int32_t dvbapi_stop_filternum(int32_t demux_index, int32_t num);
+int32_t dvbapi_stop_filter(int32_t demux_index, int32_t type);
+struct s_dvbapi_priority *dvbapi_check_prio_match(int32_t demux_id, int32_t pidindex, char type);
 
 #ifdef WITH_STAPI
 	#define cs_log(x...)	cs_log("stapi: "x)
